test: 合并 router 元数据测试并清理重复
- 合并 test_general/general_extended/endpoints/logs_extended 为 test_api_metadata.py（4 个重复断言去重）
- 删除 test_sync_extended.py（6 例被 test_sync.py 覆盖）
- test_config_llm.py 删除 4 个与 test_config_secret_crypto 重复的 is_sensitive 用例
